requirements.txt
ultralytics>=8.0.0
opencv-python
roboflow
matplotlib
seaborn
pandas
numpy
python-dotenv